Add optional removal of descriptions
Leaving in all descriptions from a Kubernetes schema increases the size of a
generated CRDs significantely and can lead to a CRD getting bigger than the
maximum allowed object size.  This commit adds an option to leave out
these descriptions.  Descriptions set on other fields in the original CRD
template are not removed and can still be used to document the CR.